refactor(station-manager): move unit of work to services
