+++ backend/station_manager/app/src/main/java/com/StationManager/app/services/unitofwork/IUnitOfWork.java
@@ -1,4 +1,4 @@
-package com.StationManager.app.storage.unitofwork;
+package com.StationManager.app.services.unitofwork;
 
 import com.StationManager.app.domain.events.Event;
