+++ backend/station_manager/app/src/main/java/com/StationManager/app/services/unitofwork/InMemoryUnitOfWork.java
@@ -1,34 +1,20 @@
-package com.StationManager.app.storage.unitofwork;
+package com.StationManager.app.services.unitofwork;
 
-import com.StationManager.app.domain.client.*;
 import com.StationManager.app.domain.events.Event;
-import com.StationManager.app.domain.train_station.Hall;
-import com.StationManager.app.domain.train_station.TicketOffice;
-import com.StationManager.app.domain.train_station.TrainStation;
 import com.StationManager.app.storage.repository.*;
 import com.StationManager.app.storage.repository.inmemory.*;
-import com.StationManager.app.storage.repository.postgres.PostgresRepository;
-import org.hibernate.Session;
-import org.hibernate.cfg.Configuration;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class PostgresUnitOfWork extends UnitOfWork {
-    public Session session;
-//
+public class InMemoryUnitOfWork extends UnitOfWork {
     public final IClientRepository clientRepository;
     public final IHallRepository hallRepository;
     public final IPrivilegyRepository privilegyRepository;
     public final ITicketOfficeRepository ticketOfficeRepository;
     public final ITrainStationRepository trainStationRepository;
 
-    public PostgresUnitOfWork() {
-        // this.session = PgSessionFactory.openSession();  // TODO: replace with real session factory
-        //noinspection resource - session is closed in close() method
-        this.session = new Configuration().configure().buildSessionFactory().openSession();
-
-        // replace wiht correct repos
+    public InMemoryUnitOfWork() {
         this.clientRepository = new InMemoryClientRepository();
         this.hallRepository = new InMemoryHallRepository();
         this.privilegyRepository = new InMemoryPrivilegyRepository();
@@ -54,19 +40,18 @@ public class PostgresUnitOfWork extends UnitOfWork {
 
     @Override
     public void commit() {
-        session.getTransaction().commit();
+        // not required for in memory storage
         logger.info("Changes committed.");
     }
 
     @Override
     public void rollback() {
-        session.getTransaction().rollback();
+        // not required for in memory storage
         logger.info("Changes rollback.");
     }
 
     @Override
     public void close() throws Exception {
         this.rollback();
-        this.session.close();
     }
 }
