+++ backend/station_manager/app/src/main/java/com/StationManager/app/services/unitofwork/UnitOfWork.java
@@ -1,0 +1,8 @@
+package com.StationManager.app.services.unitofwork;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public abstract class UnitOfWork implements IUnitOfWork {
+    protected static final Logger logger = LoggerFactory.getLogger(UnitOfWork.class);
+}
